Fix the renderer
