+++ experimental/beats/src/App.tsx
@@ -2,6 +2,7 @@ import { RouterProvider } from "@tanstack/react-router";
 import React, { useEffect } from "react";
 import { createRoot } from "react-dom/client";
 import { syncWithLocalTheme } from "./actions/theme";
+import { ImageProvider } from "./actions/ws/providers/ImageProvider";
 import { PeripheralEventsProvider } from "./actions/ws/providers/PeripheralEventsProvider";
 import { PeripheralProvider } from "./actions/ws/providers/PeripheralProvider";
 import { WSProvider } from "./actions/ws/websocket";
@@ -21,7 +22,9 @@ root.render(
     <React.StrictMode>
         <PeripheralEventsProvider>
             <PeripheralProvider>
-              <App/>
+              <ImageProvider>
+                <App/>
+              </ImageProvider>
             </PeripheralProvider>
         </PeripheralEventsProvider>
     </React.StrictMode>
